Disable policy endpoints
This PR pulls out the endpoint/view functionality of the Policy model, while
still leaving the underlying model available to use under the hood. These changes
will depend on changes being made to the group endpoints/logic to add roles to
groups through policies. Without those changes, this will break behavior in that
roles will no longer be able to be associated with users.

This approach disables the view logic and specs, with the understanding that we
may need to surface policies in the future. We may want to consider completely
pulling the logic out altogether.

The decision was made to not version the API at this point, as the concept of a
policy (in the way it's been used to this point) will be going away completely.
This will be part of the first platform-standardized release of RBAC.

- Remove policy refs from openapi.json
- Skip policy tests
- Update tests which rely on policy creation
- Remove URL definitions for policy
- Remove apidocs comments on policy