Return roles in `GET /groups/UUID/`
This adds a `roles` collection to the GET request of `/groups/UUID`, which is
generated in a custom `SerializerMethodField` by getting the policy IDs for the
group, and querying the unique roles for those policies and serializing them
with the existing `RoleMinimumSerializer`.

This is to support the management of roles via groups, while still using the
underlying policy data model.