Fix `add_fields=groups_in` filtering
Currently, when the `groups_in` param value is used on roles: `/api/rbac/v1/roles/?add_fields=groups_in`,
the `platform_default` group is always returned, regardless of whether or not
it's in the group.

We get a union of custom groups and the default: https://github.com/RedHatInsights/insights-rbac/blob/454771918c1ced6800390b712ee2e21e2b174261/rbac/management/role/serializer.py#L297-L306
but should only add the default group when the role is in the default policy.