Add concurrent index on `Tenant#ready`
This adds an index to the `Tenant` model on the `ready` column, applied concurrently [1]
so that we avoid locking the table while the index is created.

In order to accomplish this, we need to modify the migration which Django creates
by default (which is atomic) with two updates:

1) Make the migration non-atomic [2].
2) Use `AddIndexConcurrently` in the migration explicitly to ensure it's created concurrently

The reverse of this migration does drop the index.

[1] https://docs.djangoproject.com/en/4.2/ref/contrib/postgres/operations/#concurrent-index-operations
[2] https://docs.djangoproject.com/en/4.2/howto/writing-migrations/#non-atomic-migrations